+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## Unreleased changes
 
 - Refactor with extensible-0.6
+- Refactor with [elmap.hs](https://github.com/matsubara0507/elmap.hs)
+    - replace `Language.Elm` module